moved classes to each modules
