+++ repository-appuser/src/main/java/com/freshdigitable/yttt/data/source/local/AndroidPreferencesDataStore.kt
@@ -5,7 +5,6 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.intPreferencesKey
-import androidx.datastore.preferences.core.stringPreferencesKey
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -33,61 +32,5 @@ class AndroidPreferencesDataStore @Inject constructor(
     companion object {
         private val DS_CHANGE_DATE = intPreferencesKey("timeToChangeDate")
         private val DS_IS_INIT = booleanPreferencesKey("isInit")
-    }
-}
-
-@Singleton
-class YouTubeAccountAndroidDataSource @Inject constructor(
-    private val dataStore: DataStore<Preferences>,
-) {
-    val googleAccountName: Flow<String?> = dataStore.data
-        .map { it[DS_ACCOUNT_NAME] }
-
-    suspend fun putAccount(account: String) {
-        dataStore.edit {
-            it[DS_ACCOUNT_NAME] = account
-        }
-    }
-
-    companion object {
-        private val DS_ACCOUNT_NAME = stringPreferencesKey("accountName")
-    }
-}
-
-@Singleton
-class TwitchAccountAndroidDataSource @Inject constructor(
-    private val dataStore: DataStore<Preferences>,
-) {
-    val twitchToken: Flow<String?> = dataStore.data
-        .map { it[DS_TWITCH_TOKEN] }
-
-    suspend fun putTwitchToken(token: String) {
-        dataStore.edit {
-            it[DS_TWITCH_TOKEN] = token
-        }
-    }
-
-    val twitchOauthState: Flow<String?> = dataStore.data.map { it[DS_TWITCH_STATE] }
-    suspend fun putTwitchOauthState(value: String) {
-        dataStore.edit { it[DS_TWITCH_STATE] = value }
-    }
-
-    suspend fun clearTwitchOauthState() {
-        dataStore.edit { it.remove(DS_TWITCH_STATE) }
-    }
-
-    val twitchOauthStatus: Flow<String?> = dataStore.data.map { it[DS_TWITCH_STATUS] }
-    suspend fun putTwitchOauthStatus(value: String) {
-        dataStore.edit { it[DS_TWITCH_STATUS] = value }
-    }
-
-    suspend fun clearTwitchOauthStatus() {
-        dataStore.edit { it.remove(DS_TWITCH_STATUS) }
-    }
-
-    companion object {
-        private val DS_TWITCH_TOKEN = stringPreferencesKey("twitchToken")
-        private val DS_TWITCH_STATE = stringPreferencesKey("twitchOauthState")
-        private val DS_TWITCH_STATUS = stringPreferencesKey("twitchOauthStatus")
     }
 }
