+++ repository-appuser/src/main/java/com/freshdigitable/yttt/di/AccountDataStoreModule.kt
@@ -7,11 +7,6 @@ import androidx.datastore.preferences.SharedPreferencesMigration
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStoreFile
-import com.freshdigitable.yttt.data.source.TwitchAccountDataStore
-import com.freshdigitable.yttt.data.source.YouTubeAccountDataStore
-import com.freshdigitable.yttt.data.source.local.TwitchAccountLocalDataSource
-import com.freshdigitable.yttt.data.source.local.YouTubeAccountLocalDataSource
-import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -47,10 +42,4 @@ internal interface AccountDataStoreModule {
         fun provideIoCoroutineScope(): CoroutineScope =
             CoroutineScope(Dispatchers.IO + SupervisorJob())
     }
-
-    @Binds
-    fun bindYouTubeAccountLocalDataSource(source: YouTubeAccountLocalDataSource): YouTubeAccountDataStore.Local
-
-    @Binds
-    fun bindTwitchAccountLocalDataSource(source: TwitchAccountLocalDataSource): TwitchAccountDataStore.Local
 }
