+++ repository-twitch/src/main/java/com/freshdigitable/yttt/data/TwitchAccountRepository.kt
@@ -1,0 +1,10 @@
+package com.freshdigitable.yttt.data
+
+import com.freshdigitable.yttt.data.source.TwitchAccountDataStore
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class TwitchAccountRepository @Inject constructor(
+    local: TwitchAccountDataStore.Local
+) : TwitchAccountDataStore by local
