+++ repository-twitch/src/main/java/com/freshdigitable/yttt/data/source/TwitchAccountDataStore.kt
@@ -2,15 +2,6 @@ package com.freshdigitable.yttt.data.source
 
 import kotlinx.coroutines.flow.Flow
 
-interface YouTubeAccountDataStore {
-    val googleAccount: Flow<String?>
-    fun getAccount(): String?
-    suspend fun putAccount(account: String)
-    fun hasAccount(): Boolean = getAccount() != null
-
-    interface Local : YouTubeAccountDataStore
-}
-
 interface TwitchAccountDataStore {
     val twitchToken: Flow<String?>
     fun getTwitchToken(): String?
