+++ repository-youtube/build.gradle.kts
@@ -46,6 +46,7 @@ dependencies {
     implementation(libs.play.services.auth)
     implementation(libs.hilt.android)
     kapt(libs.hilt.compiler)
+    implementation(libs.androidx.datastore.preferences)
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
