+++ repository-youtube/src/main/java/com/freshdigitable/yttt/data/YouTubeAccountRepository.kt
@@ -1,6 +1,5 @@
 package com.freshdigitable.yttt.data
 
-import com.freshdigitable.yttt.data.source.TwitchAccountDataStore
 import com.freshdigitable.yttt.data.source.YouTubeAccountDataStore
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -9,7 +8,3 @@ import javax.inject.Singleton
 class YouTubeAccountRepository @Inject constructor(
     local: YouTubeAccountDataStore.Local
 ) : YouTubeAccountDataStore by local
-
-class TwitchAccountRepository @Inject constructor(
-    local: TwitchAccountDataStore.Local
-) : TwitchAccountDataStore by local
