+++ repository-youtube/src/main/java/com/freshdigitable/yttt/data/source/YouTubeAccountDataStore.kt
@@ -1,0 +1,12 @@
+package com.freshdigitable.yttt.data.source
+
+import kotlinx.coroutines.flow.Flow
+
+interface YouTubeAccountDataStore {
+    val googleAccount: Flow<String?>
+    fun getAccount(): String?
+    suspend fun putAccount(account: String)
+    fun hasAccount(): Boolean = getAccount() != null
+
+    interface Local : YouTubeAccountDataStore
+}
